Update Checkout Component Have Outline of Design
Update the Checkout component to outline the design as a starting point
of making sure that the user, if they come from directly to the
checkout, will not error out.

Also, to show where the forum would go, that creates a component for the
forum, thus splitting the page's logic from the forum and making it more
portable.