Move Purchase Action From Spot to Checkout Actions
Move the Purchase action from Spot to Checkout actions to keep all of
the actions meant for the Checkout reducer together. Thus, it is easier
to move the Checkout component between projects and allow the developer
to know which action belongs to which reducers and reduce the context
they have to know before beginning.

Furthermore, suppose the developer does not need to take a few hours to
learn how an application works. In that case, they can quickly pick up
any part of the application and add features or fix bugs much quicker
and improve their productivity.