Add Field Common Component
Add a common component called Field, which is for the forums using
Formik, so that there is a single component that does all of the
validation and has the styling, thus keep the same look and feel throw
out the application. They have all of the essential elements in one spot
to update them both functionally and style-wise quickly.