Add Spot Details Component to Search Component
Add the Spot Details component to the Search component to show more
details of the spot and allow the user to book a parking spot. As a
result of having a Spot Details component now, the Spot List component's
link has changed to display the Details component, and then the user can
purchase a parking spot.