Add Animation When Display Spot Details Component
Add animation when displaying the Spot Details component so that when
the component appears or disappears, it does it gracefully and present
to the eyes.

The framer-motion components are outside the Spot Details component
since they need an instance of the Spot Details component to do the
animation effect on it. If the framer-motion components are inside the
components, they will manipulate the component to achieve that
animation effect.

---

Ref:

https://www.framer.com/api/motion/animate-presence/#usage