Add Test Helper File For Common Helper Funciton
Add a test helper file for standard helper function such as getting
around the ESLint rule where only three callback nested are allowed.
However, this is not correct as it makes the code more complicated since
it does not consider the callback's size into consideration.
The callbacks for `waitFor` are generally small and easy to read at a
glance. These callbacks; should be excluded from the number of
callbacks, which is for callbacks that are large in comparison.