Add Integration Test For Buying a Spot
Add an integration test for Buying a Spot so that it is we know that the
whole process works as expected. As a result that the Buying a Spot
integration is so similar to submitting the form. It has merged with
Buying a Spot to keep the testing complexity to a minimum.